[feat] #1 LogInScreen snackbar 구현
