+++ app/src/main/java/org/sopt/and/LogInActivity.kt
@@ -1,22 +1,19 @@
 package org.sopt.and
 
-import android.app.Activity
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.ui.Modifier
+import org.sopt.and.screen.LogInScreen
 import org.sopt.and.ui.theme.ANDANDROIDTheme
 
-class SignInActivity : ComponentActivity() {
+class LogInActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
 
 
         super.onCreate(savedInstanceState)
         setContent {
             ANDANDROIDTheme {
-                SignInScreen()
+                    LogInScreen()
             }
         }
     }
