+++ app/src/main/java/org/sopt/and/MainActivity.kt
@@ -12,7 +12,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             ANDANDROIDTheme {
-                Intent(this, SignInActivity::class.java).apply {
+                Intent(this, LogInActivity::class.java).apply {
                     flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
                     startActivity(this)
                     finish()
